// src/data/types.ts
export interface Dyne {
  id: string;
  name: string;
  fillType: string;
  warmth: string;
  dyneType: string;
  dimensions: string;
  fillWeight: number;
  season: string;
  coolingEffect: string;
  luxury: string;
  weightPreference: string;
  garanti: string;

  image: string;
  description: string;

}